feat: improve pipeline logging and fix lat/lon schema
- Add data volume logging (MiB) for BigQuery fetches and writes
- Add pipeline duration timing (seconds and minutes)
- Add peak memory usage logging
- Change dst_lat/dst_lon from FLOAT64 to INT64 in schema
  (values are always integers - multiples of 20 for bucketing)

+++ sql/01_create_indexer_scores_table.sql
@@ -44,9 +44,10 @@ CREATE TABLE IF NOT EXISTS `graph-mainnet.iisa_data_for_dips.indexer_scores` (
   indexing_agreement_acceptance_latency FLOAT64,
 
   -- Organization/location for decentralization checks
+  -- Lat/lon are rounded to nearest 20 degrees for privacy bucketing
   org STRING,
-  dst_lat FLOAT64,
-  dst_lon FLOAT64,
+  dst_lat INT64,
+  dst_lon INT64,
 
   -- Metadata
   computed_at TIMESTAMP NOT NULL,
